DRAW-198 refactor: ws messenger 로직 리팩터링
